http.handlers: start_receiving(): Be careful before using stream
In a previous commit we separated attach() and start_reading(), and
so we cannot assume that start_reading() will be invoked immediately
after attach().  This is, for example, the case of http.protocols,
where start_reading() is delayed because we need to wait for other
constructors to complete.  So, if there is a delay between attach()
and start_reading(), it's possible that the peer closes the TCP str-
eam, and, in this case, self.stream becomes None.  For this reason
we must be careful in start_reading(), and we must check for stream
to be not None before trying to use it.