http.utils: make_filename(): Make sure we don't loop forever
Stop after 100 attempts if it's not possible to generate an unique
output file name, rather than running "forever".  This is consistent
with ftp(1)'s behavior under OpenBSD (and possily other BSDs).

Yes, the check might be paranoid, because it's very unlikely that
we have, say, index.html ... index.html.99 in the current working
directory--but I think that it's wiser to provide a stop rule.