statusicon: Add support for command-line options
By default neubot puts a visible status icon in the notification
area only when a transmission test is in progress; but the user
might want to have a persistent icon (option -n), or that the icon
blinks during the test (option -B).  While at it, add also the
standard -V and -v options, and let the user override the default
address and port.