ui: UIServer: Simple javascript support for tracking state
This patch adds neubot simple javascript support for tracking the
value of the state of the program.  The general idea is that we
append <header> a couple of <script> tags, one for jquery.js and
the other for state.js.  The former is a copy of the well-known
javascript library, and the other is part of our custom javascript
library (based on jquery).  The code in state.js periodically
sends AJAX requests to /state/change (that employs comet) to be
notified of each state change.  When the server responds (due to
changed state or timeout) state.js modifies the <span> of the
HTML page that contains the previous value of the state, and this
is possible because now we add the attribute 'id="state"' to the
<span> that contains the state value.

So, more in detail:

  1. In UIServer.get_request() dispatch control to _do_javascript()
     if the request URI starts with '/js/'.

  2. If the request is for jquery.js, _do_javascript() redirects to
     Google API.

  3. If the request is for state.js, _do_javascript() generates on-
     the-fly a piece of javascript that starts the AJAX request, and
     modifies the current page according to the response.

  4. When we generate HTML pages for / and /state, add <script> tags
     to <headers>, and so the client case use our Javascripts.

  5. When we generate HTML pages for / and /state, add the attribute
     'id="state"' to the <span> that contains the state value.

  6. Modify XHTMLComposer and DivContent to allow for #4 and #5.