Makefile: make release: Do some work to update apt repository
Move the line that creates source archives below the code block
that creates the debian package and helps in updating APT because
the process of updating APT fails if we don't move such line.

This is strange -- and I have not investigated it -- but probably
it has something to do with dpkg-scanpackages and the presence of
plain vanilla source packages.  Does dpkg-scanpackages assume they
are, instead, debian source packages?