network: Remove unused buffers module
The technique of sendbuff was good and has already been
implemented for example in Handler.  The one of recvbuff
was not very good and the reason is that it was invoking
its vent() method too frequently.  Also it was not such a
good idea to bufferize, say, a body piece, and then to pass
it to the higher layers.  The code in Handler has a better
approach because it tries to minimize the amount of bytes
that we bufferize.