log: We don't want access-logs to be part of /logs
Otherwise each client request for /logs will cause a new log entry
to be written, and that's not sane: each comet request for state
updates would cause a log entry to be written and such new log entry
would in turn cause a state update.  So, a quick solution is this
patch, that special-cases access-logs but makes not possible for the
web-interface user to read them.  In the long term we might want to
improve the code so that the user has a chance of reading access
logs too.