net.pollers: sched(): Bugfix scheduling algorithm
We must sum ticks() and delta in sched() and not in update_tasks()
because otherwise, when we're idle the task will be scheduled
self.timeout seconds in the future (because between sched() and
update_tasks there is a self.timeout seconds delay, caused by
select()).

For now, we don't address the issue of deleting a task that has
been registered but not appended to the task list yet, and we just
print a warning.  But we will follow-up soon.