net: Add support for group-of-streams stats
We want to be able to group certain streams and measure the upload
and download speed of such group only.  We need to do that in the
speedtest, because, when we employ more than one connection, we don't
want the result to contain the stats of each individual connection,
but the aggregate.

So, now in 0.2.4 we add support for groups of streams, and, later,
in 0.2.5 we will leverage this support and modify the code of the
speedtest.