net.pollers: Prepare SimpleStats to replace Timer
Reading the code I noticed that SimpleStats and Timer are very
similar, and so it is better to merge these two classes into a
single class.  The candidate is SimpleStats because it lives in
the poller, which is at a lower level than Timer (implemented
in http.clients).  So, with this patch we merge Timer features
into SimpleStats; then we will follow-up, rewriting Timer to be
just an alias for SimplerStats; finally, in the next release,
we will remove Timer from the tree.