net.pollers: Better solution for 'delete while pending'
Here's a better solution for the 'delete while pending' problem
caused by the previous patch that fixed the scheduling algorithm.
Here's the details:

1. The point of using a set of pending tasks and storing tuples
   inside it was that it was fast to search for the tuple in case
   of 'delete while pending'.  But now we can't do that anymore
   because we need to store the absolute time rather than the
   delta and we don't know such absolute time when we delete.

   So in case of 'delete while pending' we need to look-up the
   pending tasks to search for the right function to delete and
   it becomes pointless to use a set.

   For this reason we now use a vector.

2. So it becomes pointless to put a tuple into the pending tasks
   queue too, and it seems better to put a PollerTask object into
   the queue, instead.

3. In case of 'delete while pending' we don't remove from the
   pending queue but rather we set the func member to None and
   the time member to -1.  We also need to modify the code that
   pulls tasks from the pending queue to skip tasks having None
   func and -1 time.

BTW, authoring this patch I noticed that throughout the scheduling
code we use the task func() as a primary key and this might cause
problems when we registed twice a function that is not bound to any
class instance.  For this reason, I have put a big XXX comment in
front of the scheduling code.