speedtest: Speed-adaptive download/upload procedure
The download procedure now requests a portion of the /download file
using 'range:' and retries the download doubling the requested range
until the download time is bigger than one second.  So we are sure
that the download was not too short but at the same time we are sure
that we have not monopolized the broadband connection for too much
time.

The upload procedure still needs to be refined--so far it uploads
a body that is 1/4 of the size of the latest downloaded body.  This
is just a stupid heuristic and in the next release we will implement
something similar to what the download procedure already does.