log: Add support for pretty-printing progress
This patch modifies the log to add support for pretty-printing
progress information.  The pretty-print will look like:

  Operation in progress... done

It make sense to pretty-print when (a) we are not running in verbose
mode (because otherwise there might be many debug messages between
the initial and the final message); (b) logs are directed to the
standard error and the standard error is not redirected to a file
(because it does not make sense to pretty-print if there is not an
human reading the terminal ouput).  So, if (a) is False or (b) is
False, the output looks like:

  Operation in progress...
   [snip -- here we might have many debugging messages]
  Operation: done