speedtest: Print information on the operation in progress
Leverage the support for pretty-printing progress information we
commited into the logger to notify the user the current stage of
the speed-test.

We don't print progress bars because in most cases each stage is
expected to last not much more than one second--even if there are
cases when the stage takes more time, think at the user starting
a download while the stage is in progress.