http.handlers: Better protection vs. comet-after-close
The general idea of this patch is that of protecting public methods
of Handlers vs. comet-after-close.  We define comet-after-close a
situation when there is a pending comet response and the client has
already closed the connection.  When the client closes the connection
the close() method of the handler is invoked, but the handler itself
is not deleted because it is referenced by the code that handles the
comet response (notify.py).  So we must be prepared for any public
method of Handler to be called when the underlying connection has
already been closed.

In detail:

1. Protect passiveclose() checking self.isclosed, which is True
   when the underlying connection has been closed.

2. Protect attach() checking self.isclosed as well.

3. Improve protection for start_receiving() checking self.isclosed
   instead of self.stream for consistency with other public funcs.

4. In got_data() return instead of breaking out of the loop because
   there is no reason to break (and so it is more robust to return).

5. Do not check for self.stream out of the loop because of #4.