statusicon: Rewritten using httplib and threading
The previous version of this module had the problem that it is
not very simple to integrate Gtk's loop and neubot's loop in a
portable way (read: in a way that works on Windows too).

So I am rewriting it, and this time I employ httplib and threading
rather than neubot.http and non-blocking I/O.

The reason why the thread that tracks the state employs httplib
rather than neubot.http is pratical: My everyday system ships with
pygtk compiled for python 2.5 and (i) it's not pratical for me to
recompile pygtk, and (ii) neubot is not yet ready for running on
python 2.5.