statusicon: Initial implementation of the popup menu
I have added two entries: one to close the statusicon; the other
to launch neubot's web interface into a browser.  I use webbrowser
for that--but there is an issue: In some cases it's possible for
the webbrowser.open() call to block, depending on the best browser
it picks for the platform.  So I should follow-up with checks to
be sure that the call is not going to hang the status-icon.