statusicon: Ubuntu has deprecated status-icon usage
While looking for documentation on the status-icon I've read a
blog entry on design.canonical.com that explains why Ubuntu is
going to zap the notification area.  Their arguments strikes me
as very reasonable and so I have included them into a comment at
the beginning of the file.  This does not mean that I want to get
rid of the status-icon; I think that neubot should keep the icon
but it should also provide other (better?) mechanisms.