Makefile: Update commented-out rules to create installer using Wine
I am working on this issue because it would be just great for me
to be able to make a release without starting Windows up.  But there
are still some issues that I need to address and so, while I push
an update, I leave the code commented for now.