Globally remove deprecated register_func()
register_func(), register_periodic() and unregister_periodic() have
been deprecated for a long time, and it would be nice to remove them
because the poller has to be a little more complex than needed to
emulate their behavior.  Here we start with register_func() which is
simpler to remove, and we'll follow-up with the other functions in
a release or two.