net.streams: Avoid pychecker warning
Pychecker was complaining because self.readable and self.writable are
defined and used as attributes in Stream but inherited as methods from
Pollable.  To avoid this warning, we define self.handleReadable and
self.handleWritable in Stream, and we invoke, respectively, the former
from self.readable() and the latter from self.writable().