notify: Add RENEGOTIATE event for speedtest
The next version of the speedtest module needs an event named like
this which is triggered whenever the client performing the test has
finished (or after a given timeout).  In response to this event the
speedtest module will update the waiting clients about their queue
position, and, in particular, will also unchoke the first one.