notify: Attach a timestamp to each event
We attach the program state the timestamp when the state itself last
changed and send this information to the client.  The client processes
the XML message and then sends another GET /api/state request, specifying
the timestamp of the last update.  Then the server will check whether
the state is changed since the timestamp the client passed.  If so,
it will immediately send the client the new copy of the state (needless
to say, timestamped).  Otherwise the server will wait until a new state
change occurs, and then sends the new state to the client.

This amendment makes it possible to track very fast transitions that
happens between the receipit of the response and the send of the next
request (for example it was common to miss the state transition between
collect and sleeping because collect was very quick).