database: Make sure connection always exists
When we invoke sqlite3.connect() we're invoked by the initializer
of DatabaseManager, __init__().  So, if sqlite3.connect() fails the
destructor __del__() of DatbaseManager will immediately be invoked.
But, in this case, self.connection will not exist because it is
created only if sqlite3.connect() is successful.  And so, when the
destructor runs, there is an exception, which we can avoid if we
initialize self.connection to None before trying to connect to the
database.  This patch does precisely that.