Makefile: make release: put the same content in deb and tarball
In case of `make release':

Make sure we put the same content (the content of HEAD) into the deb and
tarball. Before this patch the content of the tarball was up to date with
the latest tag, while the content of the the deb was up to date with HEAD.

While at it, remove the _archive command and invoke git-archive directly
in the `release' target.