main: Without arguments try to do the right thing
If the user provided no arguments check whether an instance of
neubot is already running. If not, start a background instance of
neubot. When we know that neubot is running in background, open
a web browser (if possible) and point it to the address where the
background instance is listening.