ui: add code to track the state of the local neubot daemon
The code is split into two parts: there is a generic class that
track the state, and a more specific class that prints on the standard
output the evolution of the state.

The reason why there are two classes is that the generic one is
needed also to implement the StatusIcon and possibly other consumers,
for example I think it might make sense to implement a module that
use PyNotify and notifies the evolution of the state.