statusicon: by default detach from shell and run in background
More in detail: added the -d command line options that prevents the
statusicon to detach from the shell and run in background.  This could
be useful for debugging but in the common case one just wants the icon
to pop in the notification area (so the default is to background and
the foreground is an option).