ui: print error on 404 rather than debug message always
Rationale: in the common case it's not interesting to know the
normalize path, while it is very interesting to know it when the
daemon is in background and it cannot find a static file.  And
an error is printed every time there is an error so we just need
to read the log (instead of starting neubot again in from console
in debug mode).