move testing code into neubot.net and remove testing/
testing was once a module of neubot than we moved it outside of
neubot/ and we maintained it for a while. then the development
continued and we added unit testing to most modules at the point
that it's clear that the style of neubot is that of adding to
each module its unit test. so, now we move the code in testing
back into neubot.net.

in particular:

1. move the content of discard.py echo.py source.py into
   neubot/net/test.py (we cannot move them into streams.py
   because of circular dependencies);

2. move (and comment out) the code in utils.py into the
   streams.py module (the code is commented out because it
   might be interesting but at the moment it's not needed
   in neubot);