database: support for identifying each neubot with a random UUID
This patch adds neubot/database.py support for identifying each neubot
with a random UUID (universal unique identifier).  This identifier is
sent along when neubot uploads results of a transmission test to the
server you performed such test with.  The purpose of this identifier is
to add information useful for data analysis, e.g. it would be possible
to review the measurement history of a given neubot looking for certain
patterns, such as the connection speed decreasing near the end of the
month because the user exceeded a bandwidth cap.

With respect to privacy, we believe that the impact of adding such unique
identifier is negligible.  Indeed, neubot DOES NOT monitor user activity
but rather actively probes the network generating synthetic traffic, and
the only personal data it collects is the internet address.  So, adding
such unique identifier, we would at most be able to say that somebody
running instance XYZ of neubot has changed its internet address, and
thus, possibly, geographic location and/or provider (and note that this
information is functional to our goal of building a network neutrality
map organized by geographic location and provider).

In conclusion, we strongly believe that you should not be concerned by
this mechanism.  However, should you still be VERY worried by this issue,
you can still periodically force neubot to generate and use A NEW random
unique identifier.  You just need to become root, stop neubot, and then
run the following command:

  neubot database -f

at this point you can restart neubot.

[Many thanks to my colleague Silvia Bisi that reviewed this commit
 message and provided useful feedback. Thank you!]