neubot.nsi: roll-out cleaner uninstall and reinstall
When we reinstall, make sure that the uninstaller exists and then
invoke it (that's better than blindly invoking a program that might
not even exist :-P).

When we uninstall or reinstall wait for a while after execwait for
the system to release the locks on the files we want to (respectively)
remove or overwrite. Indeed tests on the field suggests that there
is a short time frame after execwait returns when Windows still holds
a lock on, e.g. neubot.exe, uninstall.exe.