Add menu entry to show status icon in the notification area
I should note that the way to go is to create a gnomeapplet rather
than a notification area icon (in particular because the .desktop file
installs a persistent icon in the notification area, which is quite
not the notification area purpose, btw).

However, since I've not yet written a gnomeapplet (though I've begin
reading a tutorial) for now we will be very happy with the statusicon
and eventually, someday, we'll replace the icon with a gnomeapplet.