release.sh: by default make a new patch release
I've changed my mind with respect to numbering and updated release.sh
accordingly.  The old rule was that 0.3 comes after 0.2.9, and the new
rule is that 0.2.10 comes after 0.2.9.  The rationale is that we aim to
make a release every week ("release early, release often") and prefer
to be free to make as much patch releases as we like.

So, now release.sh makes by default a patch release.  If you want to
make a new major/minor release you need to pass the release number as
the first argument, e.g. ./release.sh 1.0.0