neubot.nsi: remember to autostart neubot
This commit is very important because we don't want to bother
Windows users at the point that they need to remember to start
neubot every time they login.

BTW an interesting question is: how we notify users that there
is a new version of neubot runs in background and the user never
opens neubot web interface? Should we open the browser or popup
a window or what?