net.connectors: propagate error if connect() times out
Politecnico di Torino wireless network drops outgoing SYN segments
directed to port 9773, and so rendezvous fails because connection
times-out.

net/connectors.py is buggy because, in case of timeout, it does not
notify the controlling class that connect() failed.

So, neubot state stays "rendezvous" forever when a connection attempt
to master.neubot.org:9773 fails due to timeout.

This patch amends net/connectors.py and makes sure that, in case of
timeout, it notifies the controlling class.

Still, the rendezvous fails from Politecnico di Torino wireless
network because each connection attempt times out.  Thus, it might
make sense to move rendezvous away from port 9773 and to employ
a port which is less likely to be filtered.