[macosx] log: use syslog wrapper instead of SysLogHandler
SysLogHandler uses UDP by default.  Or you can configure it to
use a given UNIX domain socket.  The problem is that the socket
is /dev/log for most Unices, but not for MacOS X.

Then, if we use SysLogHandler we have to know the path where
the UNIX domain socket is.  A more portable solution is to use
syslog wrapper.  Indeed, who wrote syslog for the current system
of course knows the mechanism employed (i.e., UDP or UNIX domain
socket), paths, and stuff.

In the long term we want to use syslog wrapper for all Unices,
but, for now, let's apply the new behavior to MacOS X only.