[macosx] App skeleton
We support MacOS X >= 10.5 because neubot requires Python >= 2.5
and that version of MacOS is the first one to meet this requirement.
Of course, one can always install Python 2.7 on MacOS 10.4 and
adjust Info.plist and symlink /usr/bin/python to Python 2.7.  By the
way that what I'm doing with my test MacOS here at NEXA (which comes
from the PPC era).