[www-misc] neubot.js: do not bring browser on its knees
Make sure that neubot is running.  Then open the web interface.  Then
stop neubot.  You will see your browser using nearly 100% of the CPU.

This is quite annoying.  This patch modifies neubot.js to avoid this
problem.  In particular, if curtime is undefined wait five seconds
before sending the next AJAX request.

BTW, there is something that I'm missing.  I expected the error handler
to be invoked in case of neubot being down.  But it seems the success
handler is invoked instead.