[fixes] database: raise the value of MAXROWS
We used to keep 4096 rows.  But this is not enough for plots.  And
the database does not eat too much disk space.  Thus, it's OK to raise
the limit to 103680 (the expected number of rows after one year of
continuous usage at a rate of one test every five minutes).