[testing] release.sh: modified code that updates ChangeLog
This is the first neubot relase with more than one author and the
output of `git shortlog' is not exactly what I want when there is
more than one author.

Instead, generate the same output as before, but using `git log'
and specifying a custom format.

By the way, this also makes the code much more simple.