neubot/speedtest.py: Hack to avoid client disconnection
http.Connection decrements self.left after each request, and closes
the socket when such number becomes zero.  SpeedtestClient, that
use http.Connection, might fail to speedtest because the Connection
might close the socket when self.left becomes zero, e.g. after a
given number of negotiations (or possibly even in the middle of
the test.)

In principle, we don't want that because this behavior prevents
some clients from receiving the service.  And this patch provides
a simple fix, forcing self.left to be always equal to 128.  But
this hackish solution should be improved.

BTW, second-thinking at this issue, I'm not sure it is a good idea
to add another client to an already huge queue.  Perhaps it might be
better to `drop` the connection with the incoming client in order
to keep the queue length under control.  We'll third-think at that.