neubot/speedtest.py: Improve queue algorithm
The main problem with the existing queue algorithm is that it is
not scalable enough, because it allows just one transmission test
at once.  This patch tries to improve the queue algorithm to support
up to three concurrent tests (actually three is just a parameter
that could be raised at will.)

The existing algorithm
``````````````````````

The existing algorithm keeps track of the clients using a (fifo)
queue that contains the unique identifier of each client.  This
identifier is an uuid4 chosen by the server and assigned to the
client.  The client MUST pass its unique identifier to the server
using the 'authorization' header.

In check_request_headers() the server checks whether the URI that
the client wants to access is restricted or not.  Basically, all
URIs are restricted but the one used to negotiate.  If the URI is
restricted and there is not the unique identifier or the unique
identifier passed by the client is not the first in the queue, the
connection is closed.  So, there is the implied assumption that
just the first client in queue is authorized to take the test.

In do_negotiate() the server will generate an unique identifier for
the client, if the client did not provided one.  Then the client
will be granted the authorization if and only if it is the first
one in queue (as said above, the first one is the authorized one.)

The function to calculate the position in queue of each client is
very bad, because it will scan the list for each client--and so its
complexity is roughly O(N^2).

When a client uploads the results to /speedtest/collect, the server
invokes _speedtest_complete().  This function will pop the first
element from the queue and start a new negotiation.

To avoid starvation, the function _speedtest_check_timeout() is
invoked every 3 seconds.  If it finds that the first client in queue
has not completed the test in 30 seconds, this function will pop
the first element from the queue and start a new negotiation.

This is bad because if the connection with the first client in
queue is lost, the queue will be blocked for 30 seconds (in the
worst case) because the `connection lost` event is not used to
trigger a new negotiation.

The new algorithm
`````````````````

There is a class, SessionState, for each connected client, and
this class contains the following fields:

active
  whether this client is performing a test (true) or waiting in
  queue for its turn (false).

timestamp
  the timestamp of the last time we have seen request headers
  from this client.

identifier
  this is a reference to the unique identifier.

queuepos
  the position in queue of this client.

negotiations
  the number of negotiations completed by this client.

We model the queue with _SessionTrackerMixin, a mixin that contains
the following fields:

identifiers
  map an unique identifier to the related session object.

queue
  keep track of the position in queue (fifo).

The algorithm to decide whether a client is authorized or not to
access restricted URIs is now very simple.  It has the permission
(see check_request_headers()) if the session_active() method of
the new mixin returns True.  This method returns True if the uuid
of the new sessions is known _and_ the session is active.

The negotiation algorithm now is as follows.  The function named
session_negotiate of the _SessionTrackerMixin is now invoked.  This
function returns the new session.  The new session will eventually
have its active field set to True if it's one of the first three
sessions in queue.  In this case it will be authorized to perform
the test, and otherwise it will wait in queue.

When a session uploads the results to /speedtest/collect, the
function _speedtest_complete() is invoked, as before.  Now this
function deletes the current session and then starts a new
negotiation.

Periodically the session_prune() method of the new mixin is
invoked.  This method prunes all the sessions that have been
idle for more than 30 seconds and returns True if it has
removed at least one session.  In this case a new negotiation
will take place.

Whenever the queue changes (be it because we added a new session
or we removed an old one or because we pruned stale sessions),
the state of the queue is updated (see _do_update_queue() in the
new mixin.)

Still, this algorithm does not leverages the `connection lost` event,
but this will be the subject of an upcoming patch.