neubot/speedtest.py: Use connection_lost to improve queue algorithm
Add _SessionTrackerMixin a dictionary that maps each connection to
the identifier that identifies it.  Refresh the binding each time
we receive the headers of the new request (*please note that the
first connection would be skipped using the algorithm of this
patch*).  And, when a connection is lost, try to use the identifier
to access the related session and, if the session exist then remove
it (the first connection that terminates remove the session and
all the others will find that the session does not exist.)