neubot/speedtest.py: Make sure first session's connection is not special
The algorithm that maps a connection to the connection's session unique
identifier runs when we have read just the headers.  Therefore the first
connection of a session could not be mapped because at that point the
session does not exist.  But we want to track that connection too.  In
order to do that, this patch adds a special case that register a new
connection if this is the first negotiation of the related session.