net/pollers.py: Surround stream.closing with try..except
This patch augments neubot robustness because it's possible for the
closing() function to fail and we don't want that to tear down the
program.  BTW, note that this patch treats SystemExit the same way as
KeyboardInterrupt.  A subsequent patch will modify the readable and
writable methods to treat SystemExit like KeyboardInterrupt as well.