rendezvous.py: On -S do not start a companion speedtest server
The default behavior so far has been to run the rendezvous server
and the speedtest server in the same unprivileged background process.
The rendezvous server has always started the speedtest server as a
part of its startup procedure.

This commit modifies the rendezvous server so that it does not start
the speedtest server automatically.  In other words, we want the speedtest
server to live into its own process instead of sharing the same process
with the rendezvous server.

This change makes sense for a number of reasons.  The most important one
is that we want to deploy an alternate speedtest server and we don't want
to start up a rendezvous server just to deploy such alternate speedtest
server.