speedtest.py: Add -d command line option for -S mode
Now that the speedtest server should run as a standalone process and
not in the same process of the rendezvous server, it makes sense to
amend the default behavior and drop privileges and background.  The old
behavior (not dropping privileges and foreground) is still available
via the new -d (debug) command line switch.