net/pollers.py: Don't use from..import for standard library
Using from..import for standard library might be confusing for someone
that already knows Python but is new to the project.  So, better to use
plain import for standard library and from..import for neubot own
modules.

Many thanks to Alessio Palmero that pointed out this issue with the
source code.