neubot/notify.py: safer dispatching of events
Some time ago I removed the try..except guards when dispatching events
but this might not be the right thing to do because if something goes
wrong and the called function does not handle exceptions, then we will
break out the whole event dispatching loop.

So, it's better to add again those guards and eventually catch exceptions
also in those called functions, if we need to take a particular action,
such as returning "500 Internal Server Error".