statusicon: By default show the status icon
Yes, the specs say that the notification area is for transient messages
only but many many icons are persistent and, having a persistent icon, we
provide users a simple way to monitor neubot and interact with it.

So, now the default behavior is to show the icon and there is the option
to hide it (-q).  I will keep the old switch to hide the icon (-n) around
for a few cycles, and then I will remove it.