statusicon: Improve right-click menu labels
In particular, make sure that we don't suggest implicitly or explicitly
that the web interface is neubot.  Of course, neubot is not the web interface
but users have pointed out that there is confusion, and they are right.

BTW, this is just the first step in this direction and more work needs to
be done, expecially for MacOS.