Makefile: install menu entries for all Unix-like OSes
This is a follow-up to a previous patch that installs icons for all
Unix-like operating systems.  This patch also installs menu entries for
all Unix-like operating system, so you don't need to install debian to
have menu entries for neubot.  As usual, another patch will follow-up
to cleanup the special-case code for debian.