tools/dbtool.py: Standalone tool to analyze database
The tool allows various `processors` to do interesting things with
the content of the database while we walk it.  In particular this
patch adds a processor that tries to anonymize the database using the
PyGeoIP library.

This patch contains some fixes contributed by Roberto D'Auria.

A couple of notes:

1. the legal status of the anonymized database is not clear and we
   have made a request to <http://selili.polito.it/> in order to shed
   some more light on this issue,

2. it's not clear whether the database is accurate enough.  Some times
   I appear to live in Savona even if I'm Turin,

3. as pointed out by Alessio Palmero, it might not be a good idea to
   store (just) XMLs in the database.  In his opinion it would be better
   either to drop XML and just store data into tables, or to add the
   database also tables.  The rationale seems to be the fact that users
   will certainly want to query the database more deeply that what the
   fact that we store XMLs allows.

In conclusion, this dbtool.py is place where you should expect radical
changes in the near future, because there are many open issues.