neubot/tools: Directory for standalone tools
The idea behind standalone tools is as follows.  The tools do not
depend on any piece of code in neubot, but neubot is free to import
code from these tools.  This way it's also possible to ship the
tools as standalone Python scripts ("Here's the anonymized database
and here's an useful script to do things with it") in an easy way.

In private discussions it has been pointed out by Roberto D'Auria
that for a developer it might be a bit confusing to import code
from the tools/ subdirectory ("isn't it standalone?") and that it
would be less surprising if we design, say, database.py to stand
alone and then we eventually rename it to, say, neubot-dbase.py
and "here's the usefult script ...".

I think Roberto's objection makes sense and so it's likely that
we'll merge the content of tools/ into the relevant files of neubot
sometime soon (at least we'll merge dbtool.py and database.py).

This does not mean that tools is a bad idea because it could still
host files that neubot does not import stuff from.  And, also, it
is extremely useful to have a place where we can put little pieces
of code we have written from scratch.  Sometimes new ideas pour when
you do that.