pathnames.py: MacOS: do not hardcode '/Applications/neubot ...'
At least in one case a tester has complained that you have to copy the
neubot application in /Applications or it won't work because the paths are
hardcoded.  The tester is right and here's the fix that makes the code
independent of the position in the file system.

Note that we know the launcher script Contents/MacOS/neubot always uses
absolute paths, but we have decided to added a (now redundant) call to
os.path.abspath() for robustness.