neubot/unix.py: make sure we use LOG_DAEMON
Quite frankly, I remember that it's possible to pass openlog() the
LOG_DAEMON flag for that flag to be used as default in subsequent calls
to syslog.  But probably I am wrong and, in the end, it doesn't matter
much, since this patch fixes the issue and now neubot uses LOG_DAEMON
facility.