net/pollers.py: Allow to break the loop cleanly
At least the handler of /api/exit needs that.  Generally speaking it
is a good idea for the program to be able to generate an event that
causes the main loop to break (even if there is still something to do).

A possible follow-up to this patch would be to break the loop if and
only if the variable `again` becomes False (i.e. without even inspecting
the value of readset and writeset).