rendezvous.py: Do not wait for a deterministic amount of time
Suggested by Elias.

The point here is that it is bad for clients to attempt the next
rendezvous after a fixed amount of seconds.  Indeed this could synchronize
all the neubot that are in queue at a certain amount of time if the
speedtest server is restarted.

However (if I have correctly understood the point made by Elias) it would
be wrong to wait a random amount of time choosen uniformly in a given
interval, because this will add noise to the process of sampling the network.

To reduce the noise, it would be better to decide once and forever the
random interval between each rendezvous.  And this is what this patch does.