database.py: query_results_functional(): New query API backend
Requested by Antonio.  This function allows to query the database
using timing (e.g. since..until) instead of positions in the cache.
Note that this function bypasses the cache and queries the database
directly.  The database should not be slow and I was probably in
error when I decided to use a cache.